feat: add AuthController/AuthService and update main controllers, services, entities\n\n- Added authentication flow (AuthController, AuthService, AuthResponse DTO)\n- Updated User, Post, Category, and Tag entities\n- Updated UserService, PostService, PostRepo\n- Updated controllers for User, Post, Category, Tag\n- Updated global exception handling and ErrorResponse DTO\n- Updated SecurityConfig and application.properties

+++ src/main/java/com/abdelrahman/blogplatorm/services/PostService.java
@@ -6,6 +6,10 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort.Direction;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -38,15 +42,14 @@ public class PostService {
 	private final UserRepo userRepo;	
 	private final CategoryService catService;
 
+	@Transactional
 	public PostResponseDto insert(PostRequestDto requestDto) {
 		Set<Tag> tags = new HashSet<>(tagService.getOrCreateByNames(requestDto.getTags()));
-		
-		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-		MyUserPrinciple user = (MyUserPrinciple) auth.getPrincipal();
-		if(!user.getId().equals(requestDto.getUserId())) {
+
+		if(!isCurrentUser(requestDto.getUserId())) 
 			throw new AccessDeniedException("userId Not Allowed");
-		}
-	 	User author = userRepo.findById(user.getId()).orElseThrow(()->new RecordNotFoundException("User Not Found"));
+		
+	 	User author = userRepo.findById(requestDto.getUserId()).orElseThrow(()->new RecordNotFoundException("User Not Found"));
 		Category category = catService.getById(requestDto.getCategoryId());		
 		
 		Post post = mapper.toPost(requestDto);
@@ -60,15 +63,16 @@ public class PostService {
 		return mapper.toPostDto(postRepo.save(post));
 	}
 
-	public List<PostResponseDto> findAll() {
+	public Page<PostResponseDto> findAll(int pageNumber,int pageSize,String sortBy ,String sortType) {
 		
-		return mapper.toListDto(postRepo.findAllByStatus(Status.PUBLISHED));
+		Pageable page = PageRequest.of(pageNumber, pageSize, Direction.fromString(sortType),sortBy);
+		Page<Post> posts = postRepo.findAllByStatus(Status.PUBLISHED,page);
+		return posts.map(post->mapper.toPostDto(post));
 	}
 	
 	public List<PostResponseDto> findDraftsForCurrentUser() {
-	    Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-	    MyUserPrinciple user = (MyUserPrinciple) auth.getPrincipal();
-
+	    MyUserPrinciple user = getCurrentUser();
+	    
 	    return mapper.toListDto(
 	        postRepo.findByUserIdAndStatus(user.getId(), Status.DRAFT)
 	    );
@@ -80,22 +84,26 @@ public class PostService {
 	}
 	
 	public List<PostResponseDto> findByTitle(String title) {
-		return 	mapper.toListDto(postRepo.findByTitleAndStatus(title,Status.PUBLISHED));
+		return 	mapper.toListDto(postRepo.findByTitleContainingIgnoreCaseAndStatus(title,Status.PUBLISHED));
 	}
 
-	public String delete(Long id) {
+	@Transactional
+	public void delete(Long id) {
 		if(!postRepo.existsById(id)) {
-			return "This post is not found to delete";
+			throw new RecordNotFoundException("This post is not found!");
 		}
-		else {
-			postRepo.deleteById(id);
-			return "This post deleted successfully";
-		}
+		if( !isCurrentUser(postRepo.getReferenceById(id).getUser().getId()))
+			throw  new AccessDeniedException("Not allowed");
+		postRepo.deleteById(id);
 	}
 	
 	@Transactional
 	public PostResponseDto update(Long id,PostUpdateDto postUpdate) {
+		
 		Post post = postRepo.findById(id).orElseThrow(()->new RecordNotFoundException("Post Not Found"));
+		if(!isCurrentUser(post.getUser().getId()))
+		    throw new AccessDeniedException("You are not allowed to update this post");
+		
 		if(postUpdate.getTitle()!=null && !postUpdate.getTitle().isBlank()) {
 			post.setTitle(postUpdate.getTitle());
 		}
@@ -116,15 +124,29 @@ public class PostService {
 		}
 		return mapper.toPostDto(postRepo.save(post));
 	}
-	
+	@Transactional
 	public PostResponseDto publishPost(Long id) {
 		Post post = postRepo.findById(id).orElseThrow(()->new RecordNotFoundException("Not found post"));
+		if(!isCurrentUser(post.getUser().getId()))
+		    throw new AccessDeniedException("You are not allowed to update this post");
+	
 		if(post.getStatus()==Status.PUBLISHED) {
 			throw new RuntimeException("Post already published");
 		}
 		post.setStatus(Status.PUBLISHED);
 		post.setPublishedAt(LocalDateTime.now());
 		return mapper.toPostDto(postRepo.save(post));
+	}
+	
+	private MyUserPrinciple getCurrentUser() {
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		return (MyUserPrinciple) auth.getPrincipal();
+	}
+	
+	public boolean isCurrentUser(Long userId) {
+		
+		MyUserPrinciple currentUser = getCurrentUser();
+		return currentUser.getId().equals(userId);
 	}
 
 }
